[v2.7.x] prov/efa: make the current reverse AV lock free
Replace the uthash cur_reverse_av (explicit and implicit) with an efa_av_array,
so the lookup in the cq read is lock free.

The array is indexed by a 32-bit key built from the 16-bit AHN and the 16-bit
QPN. The halves are bit interleaved (a Morton code) rather than packed as
ahn << 16 | qpn: interleaving keeps k-bit halves inside a 2k-bit index, so the
realistic case of small AHNs and small QPNs stays in the array's inline region
and neighboring pairs stay neighbors in the index space. Packing by shifting
would instead place consecutive AHNs 64K indexes apart and burn one chunk per
AHN. x86 gets the interleave as two PDEP instructions when the build targets
BMI2; everything else uses the portable shift/mask spread.

To address the whole 32-bit key space, efa_av_array's max_idx ceiling grows to
UINT32_MAX and max_idx widens to uint64_t.

(cherry picked from commit e2c27ad4bf33f395a9b6f1b7eebe0e05b160c491)

+++ prov/efa/test/gtest/efa_gtest_av_array_utils.c
@@ -6,8 +6,8 @@
 
 const int efa_test_av_array_inline_size = EFA_AV_ARRAY_INLINE_SIZE;
 const int efa_test_av_array_chunk_size = EFA_AV_ARRAY_CHUNK_SIZE;
-const int efa_test_av_array_max_idx_ceiling = EFA_AV_ARRAY_MAX_IDX_CEILING;
-const int efa_test_av_array_default_max_idx = EFA_AV_ARRAY_DEFAULT_MAX_IDX;
+const uint64_t efa_test_av_array_max_idx_ceiling = EFA_AV_ARRAY_MAX_IDX_CEILING;
+const uint64_t efa_test_av_array_default_max_idx = EFA_AV_ARRAY_DEFAULT_MAX_IDX;
 
 struct efa_av_array *efa_test_av_array_create(void)
 {
@@ -18,7 +18,7 @@ struct efa_av_array *efa_test_av_array_create(void)
 	return arr;
 }
 
-struct efa_av_array *efa_test_av_array_create_max(unsigned max_idx)
+struct efa_av_array *efa_test_av_array_create_max(uint64_t max_idx)
 {
 	struct efa_av_array *arr;
 	struct efa_av_array_attr attr = {0};
@@ -29,7 +29,7 @@ struct efa_av_array *efa_test_av_array_create_max(unsigned max_idx)
 	return arr;
 }
 
-struct efa_av_array *efa_test_av_array_create_attr(unsigned max_idx,
+struct efa_av_array *efa_test_av_array_create_attr(uint64_t max_idx,
 						   unsigned inline_size,
 						   unsigned chunk_size)
 {
